refactor: 프로젝트 이름 "jsonhelper2"에서 "JSONinja"로 변경
- 패키지명을 com.livteam.jsonhelper2에서 com.livteam.jsoninja로 변경
- 플러그인 ID 및 관련 클래스 이름 업데이트
- 새로운 플러그인 아이콘 추가
- 툴윈도우 팩토리 클래스 이름 변경

프로젝트 브랜딩을 더 명확하고 직관적인 이름으로 리브랜딩하였습니다.
