+++ gradle.properties
@@ -1,6 +1,6 @@
 # IntelliJ Platform Artifacts Repositories -> https://plugins.jetbrains.com/docs/intellij/intellij-artifacts.html
 
-pluginGroup = com.livteam.jsonhelper2
+pluginGroup = com.livteam.jsoninja
 pluginName = json-helper2
 pluginRepositoryUrl = https://github.com/buYoung/json-helper2
 # SemVer format -> https://semver.org
