+++ src/main/kotlin/com/livteam/jsoninja/ui/toolWindow/JsoninjaToolWindowFactory.kt
@@ -1,12 +1,12 @@
-package com.livteam.jsonhelper2.ui.toolWindow
+package com.livteam.jsoninja.ui.toolWindow
 
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.wm.ToolWindow
 import com.intellij.openapi.wm.ToolWindowFactory
 import com.intellij.ui.content.ContentFactory
-import com.livteam.jsonhelper2.ui.component.JsonHelperPanel
+import com.livteam.jsoninja.ui.component.JsonHelperPanel
 
-class JsonHelperToolWindowFactory : ToolWindowFactory {
+class JsoninjaToolWindowFactory : ToolWindowFactory {
     override fun createToolWindowContent(project: Project, toolWindow: ToolWindow) {
         val jsonHelperPanel = JsonHelperPanel(project)
         val content = ContentFactory.getInstance().createContent(jsonHelperPanel, "", false)
