+++ src/main/resources/META-INF/plugin.xml
@@ -1,7 +1,7 @@
 <!-- Plugin Configuration File. Read more: https://plugins.jetbrains.com/docs/intellij/plugin-configuration-file.html -->
 <idea-plugin>
-    <id>com.livteam.jsonhelper2</id>
-    <name>Json Helper2</name>
+    <id>com.livteam.jsoninja</id>
+    <name>JSONinja</name>
     <vendor>buyoung</vendor>
 
     <depends>com.intellij.modules.platform</depends>
@@ -9,11 +9,11 @@
     <resource-bundle>messages.LocalizationBundle</resource-bundle>
 
     <extensions defaultExtensionNs="com.intellij">
-        <toolWindow factoryClass="com.livteam.jsonhelper2.ui.toolWindow.JsonHelperToolWindowFactory" id="JsonHelper2"/>
+        <toolWindow factoryClass="com.livteam.jsoninja.ui.toolWindow.JsoninjaToolWindowFactory" id="JSONinja"/>
     </extensions>
 
     <applicationListeners>
-        <listener class="com.livteam.jsonhelper2.listeners.JsonHelperActivationListener" topic="com.intellij.openapi.application.ApplicationActivationListener"/>
-        <listener class="com.livteam.jsonhelper2.listeners.JsonHelperActivationListener" topic="com.intellij.ide.plugins.DynamicPluginListener"/>
+        <listener class="com.livteam.jsoninja.listeners.JsonHelperActivationListener" topic="com.intellij.openapi.application.ApplicationActivationListener"/>
+        <listener class="com.livteam.jsoninja.listeners.JsonHelperActivationListener" topic="com.intellij.ide.plugins.DynamicPluginListener"/>
     </applicationListeners>
 </idea-plugin>
