+++ src/test/kotlin/com/livteam/jsonhelper2/MyPluginTest.kt
@@ -1,4 +1,4 @@
-package com.livteam.jsonhelper2
+package com.livteam.jsoninja
 
 import com.intellij.ide.highlighter.XmlFileType
 import com.intellij.openapi.components.service
@@ -6,7 +6,7 @@ import com.intellij.psi.xml.XmlFile
 import com.intellij.testFramework.TestDataPath
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
 import com.intellij.util.PsiErrorElementUtil
-import com.livteam.jsonhelper2.services.JsonHelperProjectService
+import com.livteam.jsoninja.services.JsonHelperProjectService
 
 @TestDataPath("\$CONTENT_ROOT/src/test/testData")
 class MyPluginTest : BasePlatformTestCase() {
